+++ CHANGELOG.rst
@@ -11,7 +11,7 @@ v0.1.0 (22/04/2024)
 - Extra detail to error messages
 - Filter for invalid names
 - Item types are now displayed next to item in tree
-
+- Sample macromolecules now have their safety level displayed next to their names
 
 **Fixed**